+++ python/dalex/dalex/dataset_level/_aggregated_profiles/object.py
@@ -91,9 +91,8 @@ class AggregatedProfiles:
         self.raw_profiles = None
         self.random_state = random_state
 
-    def __str__(self):
-        from IPython.display import display
-        display(self.result)
+    def _repr_html_(self):
+        return self.result._repr_html_()
 
     def fit(self,
             ceteris_paribus,