+++ python/dalex/dalex/dataset_level/_residual_diagnostics/object.py
@@ -32,9 +32,8 @@ class ResidualDiagnostics:
         self.result = None
         self.variables = variables_
 
-    def __str__(self):
-        from IPython.display import display
-        display(self.result)
+    def _repr_html_(self):
+        return self.result._repr_html_()
 
     def fit(self, explainer):
         """Calculate the result of explanation