+++ python/dalex/dalex/dataset_level/_variable_importance/object.py
@@ -98,9 +98,8 @@ class VariableImportance:
         self.permutation = None
         self.processes = processes_
 
-    def __str__(self):
-        from IPython.display import display
-        display(self.result)
+    def _repr_html_(self):
+        return self.result._repr_html_()
 
     def fit(self, explainer):
         """Calculate the result of explanation