+++ python/dalex/dalex/instance_level/_ceteris_paribus/object.py
@@ -78,9 +78,8 @@ class CeterisParibus:
         self.new_observation = None
         self.processes = processes_
 
-    def __str__(self):
-        from IPython.display import display
-        display(self.result)
+    def _repr_html_(self):
+        return self.result._repr_html_()
 
     def fit(self,
             explainer,