+++ python/dalex/dalex/instance_level/_shap/object.py
@@ -71,9 +71,8 @@ class Shap:
         self.processes = processes_
         self.random_state = random_state_
 
-    def __str__(self):
-        from IPython.display import display
-        display(self.result)
+    def _repr_html_(self):
+        return self.result._repr_html_()
 
     def fit(self,
             explainer,